feat: create step four and five for investor
