+++ src/domain/onboarding/molecules/desktop-onboarding-step-five-investor/desktop-onboarding-step-five-investor.types.ts
@@ -1,0 +1,8 @@
+import { IInvestorDemandType } from '@interfaces';
+
+import { IDesktopOnboardingStepFiveInvestorData } from '../../onboarding.types';
+
+export interface IDesktopOnboardingStepFiveInvestorProps {
+  investorDemandTypes: IInvestorDemandType[];
+  onContinueButtonClick: (data: IDesktopOnboardingStepFiveInvestorData) => void;
+}
