+++ src/domain/onboarding/molecules/desktop-onboarding-step-four-investor/desktop-onboarding-step-four-investor.component.tsx
@@ -1,37 +1,61 @@
 import { Button, Grid } from '@mui/material';
 import { useForm } from 'react-hook-form';
 
-import { SelectWithController, TextFieldWithController } from '@ui';
+import { SelectWithController } from '@ui';
 
 import { DesktopOnboardingFormLayout } from '../../atoms';
-import { IDesktopOnboardingStepFourStartupData } from '../../onboarding.types';
+import { IDesktopOnboardingStepFourInvestorData } from '../../onboarding.types';
 import {
   mapInvestmentSizesToDropdownOptions,
   mapInvestmentStageTypesToDropdownOptions,
+  mapTeamSizesToDropdownOptions,
 } from '../../utils';
-import { IDesktopOnboardingStepFourStartupProps } from './desktop-onboarding-step-four-startup.types';
+import { IDesktopOnboardingStepFourInvestorProps } from './desktop-onboarding-step-four-investor.types';
 
-export const DesktopOnboardingStepFourStartup = ({
+export const DesktopOnboardingStepFourInvestor = ({
   investmentSizes,
   investmentStageTypes,
-}: IDesktopOnboardingStepFourStartupProps): JSX.Element => {
-  const { control, handleSubmit } = useForm<IDesktopOnboardingStepFourStartupData>({
+  teamSizes,
+}: IDesktopOnboardingStepFourInvestorProps): JSX.Element => {
+  const { control, handleSubmit } = useForm<IDesktopOnboardingStepFourInvestorData>({
     defaultValues: {
+      teamSizeIds: [],
       investmentSizeIds: [],
-      whatYouAreLookingFor: '',
       investmentStageTypeIds: [],
     },
   });
 
   const investmentStageTypesDropdownOptions =
     mapInvestmentStageTypesToDropdownOptions(investmentStageTypes);
+  const teamSizesDropdownOptions = mapTeamSizesToDropdownOptions(teamSizes);
   const investmentSizesDropdownOptions = mapInvestmentSizesToDropdownOptions(investmentSizes);
 
   const onSubmit = handleSubmit((data) => console.log(data));
 
   return (
-    <DesktopOnboardingFormLayout heading="Setup Startup" subHeading="Step 4/5" onSubmit={onSubmit}>
+    <DesktopOnboardingFormLayout heading="Setup Investor" subHeading="Step 4/5" onSubmit={onSubmit}>
       <Grid container columnSpacing={4} rowSpacing={3}>
+        <Grid item xs={12}>
+          <SelectWithController
+            controllerProps={{
+              control,
+              name: 'teamSizeIds',
+              rules: {
+                required: true,
+              },
+            }}
+            formControlProps={{
+              fullWidth: true,
+            }}
+            selectProps={{
+              multiple: true,
+              fullWidth: true,
+              label: 'Preferred Team Size',
+              options: teamSizesDropdownOptions,
+              labelId: 'desktop-onboarding-step-two-investor-team-size-id-select',
+            }}
+          />
+        </Grid>
         <Grid item xs={12}>
           <SelectWithController
             controllerProps={{
@@ -47,9 +71,9 @@ export const DesktopOnboardingStepFourStartup = ({
             selectProps={{
               multiple: true,
               fullWidth: true,
-              label: 'Investment Stage',
+              label: 'Preferred Investment Stage',
               options: investmentStageTypesDropdownOptions,
-              labelId: 'desktop-onboarding-step-four-startup-investment-stage-type-ids-select',
+              labelId: 'desktop-onboarding-step-two-investor-investment-stage-type-ids-select',
             }}
           />
         </Grid>
@@ -68,27 +92,9 @@ export const DesktopOnboardingStepFourStartup = ({
             selectProps={{
               multiple: true,
               fullWidth: true,
-              label: 'How much money do I need?',
               options: investmentSizesDropdownOptions,
-              labelId: 'desktop-onboarding-step-four-startup-investment-size-ids-select',
-            }}
-          />
-        </Grid>
-        <Grid item xs={12}>
-          <TextFieldWithController
-            controllerProps={{
-              control,
-              name: 'whatYouAreLookingFor',
-              rules: {
-                required: true,
-              },
-            }}
-            inputProps={{
-              rows: 3,
-              fullWidth: true,
-              multiline: true,
-              autoComplete: 'disabled',
-              label: 'What are you looking for?',
+              label: "How much money you'd like to give?",
+              labelId: 'desktop-onboarding-step-two-investor-investment-size-ids-select',
             }}
           />
         </Grid>
