+++ src/domain/onboarding/molecules/desktop-onboarding-step-four-investor/desktop-onboarding-step-four-investor.types.ts
@@ -1,0 +1,10 @@
+import { IInvestmentSize, IInvestmentStageType, ITeamSize } from '@interfaces';
+
+import { IDesktopOnboardingStepFourInvestorData } from '../../onboarding.types';
+
+export interface IDesktopOnboardingStepFourInvestorProps {
+  investmentSizes: IInvestmentSize[];
+  investmentStageTypes: IInvestmentStageType[];
+  onContinueButtonClick: (data: IDesktopOnboardingStepFourInvestorData) => void;
+  teamSizes: ITeamSize[];
+}
