+++ src/domain/onboarding/molecules/desktop-onboarding-step-three-investor/desktop-onboarding-step-three-investor.types.ts
@@ -1,0 +1,11 @@
+import { IFocusMarket, IIndustrialSector, IInvestorProfileType, IStartupSector } from '@interfaces';
+
+import { IDesktopOnboardingStepThreeInvestorData } from '../../onboarding.types';
+
+export interface IDesktopOnboardingStepThreeInvestorProps {
+  focusMarkets: IFocusMarket[];
+  industrialSectors: IIndustrialSector[];
+  investorProfileTypes: IInvestorProfileType[];
+  onContinueButtonClick: (data: IDesktopOnboardingStepThreeInvestorData) => void;
+  startupSectors: IStartupSector[];
+}
