+++ src/domain/onboarding/molecules/index.ts
@@ -1,6 +1,8 @@
 export * from './desktop-onboarding-step-one/desktop-onboarding-step-one.component';
 export * from './desktop-onboarding-step-two/desktop-onboarding-step-two.component';
-export * from './desktop-onboarding-step-two-investor/desktop-onboarding-step-two-investor.component';
 export * from './desktop-onboarding-step-four-startup/desktop-onboarding-step-four-startup.component';
 export * from './desktop-onboarding-step-five-startup/desktop-onboarding-step-five-startup.component';
 export * from './desktop-onboarding-step-three-startup/desktop-onboarding-step-three-startup.component';
+export * from './desktop-onboarding-step-four-investor/desktop-onboarding-step-four-investor.component';
+export * from './desktop-onboarding-step-five-investor/desktop-onboarding-step-five-investor.component';
+export * from './desktop-onboarding-step-three-investor/desktop-onboarding-step-three-investor.component';
