+++ src/domain/onboarding/utils/map-investment-stage-types-to-dropdown-options/map-investment-stage-types-to-dropdown-options.function.ts
@@ -2,7 +2,7 @@ import { ISelectOption } from '@ui';
 
 import { IInvestmentStageType } from '@interfaces';
 
-export const mapInvestmentStageTypestoDropdownOptions = (
+export const mapInvestmentStageTypesToDropdownOptions = (
   investmentStageTypes: IInvestmentStageType[],
 ): ISelectOption[] =>
   investmentStageTypes.map((_investmentStageType) => ({
