+++ src/domain/onboarding/utils/map-investment-stage-types-to-dropdown-options/map-investment-stage-types-to-dropdown-options.spec.ts
@@ -1,7 +1,7 @@
 import { functionImportTest } from '@utils';
 
-import { mapInvestmentStageTypestoDropdownOptions } from './map-investment-stage-types-to-dropdown-options.function';
+import { mapInvestmentStageTypesToDropdownOptions } from './map-investment-stage-types-to-dropdown-options.function';
 
-describe('mapInvestmentStageTypestoDropdownOptions function', () => {
-  functionImportTest(mapInvestmentStageTypestoDropdownOptions);
+describe('mapInvestmentStageTypesToDropdownOptions function', () => {
+  functionImportTest(mapInvestmentStageTypesToDropdownOptions);
 });
